chore: fix alias in chart-utils
