+++ ui/hooks/perps/index.ts
@@ -1,5 +1,5 @@
 export { usePerpsOrderForm } from './usePerpsOrderForm';
-export { usePerpsEligibility } from './usePerpsEligibility
+export { usePerpsEligibility } from './usePerpsEligibility';
 export type {
   UsePerpsOrderFormOptions,
   UsePerpsOrderFormReturn,
